write eventfd

+++ ulib/axstarry/src/syscall_fs/imp/eventfd.rs
@@ -6,7 +6,7 @@ use crate::syscall_fs::ctype::eventfd::EventFd;
 use crate::{SyscallError, SyscallResult};
 
 pub fn syscall_eventfd(args: [usize; 6]) -> SyscallResult {
-    let initval = args[0] as u32;
+    let initval = args[0] as u64;
     let flags = args[1] as i32;
     error!(
         "[syscall_eventfd] is called: initval={}, flags={}",
